move duplicate **Interpolate code into GenericInterpolate class
